% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_drift.R
\name{correct_linear}
\alias{correct_linear}
\title{Linearly Correct for QC Drift}
\usage{
correct_linear(io, abundance, qc)
}
\arguments{
\item{io}{Numeric vector indicating injection order.}

\item{abundance}{Numeric vector containing uncorrected abundances.}

\item{qc}{Logical vector identifying QC samples.}
}
\value{
A numeric vector of QC-corrected abundances.
}
\description{
This function corrects for analytical drift in mass spectrometry data
by linearly interpolating between QC samples. It assumes that the first and
last samples are QC samples.
}
