% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{generate_cv_plot}
\alias{generate_cv_plot}
\title{Generate CV Plot}
\usage{
generate_cv_plot(data, vbl, title_prefix = "Metabolite CV (QC Samples)")
}
\arguments{
\item{data}{A data frame containing metabolite abundance data.}

\item{vbl}{The name of the column in \code{data} to use for calculating the coefficient of variation.}

\item{title_prefix}{A character string to use as the prefix for the plot title.}
}
\value{
A ggplot2 object representing the CV plot.
}
\description{
Generate CV Plot
}
