% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{generate_pca_pairs}
\alias{generate_pca_pairs}
\title{Generate a pairs plot from a PCA data object}
\usage{
generate_pca_pairs(pcs, npcs = 4, qc_starts_with = "QC")
}
\arguments{
\item{pcs}{List of type \code{prcomp}}

\item{npcs}{Number of principal components to plot}

\item{qc_starts_with}{Character specifying the prefix used to identify QC samples}
}
\value{
A ggplot2 object with the pairs plot
}
\description{
Generate a pairs plot from a PCA data object
}
