% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{generate_pca_plot}
\alias{generate_pca_plot}
\title{Generate PCA Scores Plot}
\usage{
generate_pca_plot(
  pcs,
  title_prefix = "PCA",
  pc_x = 1,
  pc_y = 2,
  qc_starts_with = "QC"
)
}
\arguments{
\item{pcs}{List of type \code{prcomp}}

\item{title_prefix}{Character figure title label}

\item{pc_x}{Integer identifying the PC to be plotted on the x-axis}

\item{pc_y}{Integer identifying the PC to be plotted on the y-axis}

\item{qc_starts_with}{Character specifying the prefix used to identify QC samples}
}
\value{
A ggplot2 object representing the PCA scores plot.
}
\description{
Generate PCA Scores Plot
}
