% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{generate_qc_heatmap}
\alias{generate_qc_heatmap}
\title{Generate Heatmap}
\usage{
generate_qc_heatmap(data, vbl, title_prefix = "QC Heatmap")
}
\arguments{
\item{data}{A data frame containing metabolite abundance data.}

\item{vbl}{The name of the column in \code{data} to use for the heatmap.}

\item{title_prefix}{A character string to use as the prefix for the plot title.}
}
\value{
A ggplot2 object representing the heatmap.
}
\description{
Generate Heatmap
}
