% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{generate_qc_violin}
\alias{generate_qc_violin}
\title{Generate Violin Plot}
\usage{
generate_qc_violin(data, vbl, title_prefix = "QC Abundance Distribution")
}
\arguments{
\item{data}{A data frame containing metabolite abundance data.}

\item{vbl}{The name of the column in \code{data} to use for plotting.}

\item{title_prefix}{A character string to use as the prefix for the plot title.}
}
\value{
A ggplot2 object representing the violin plot.
}
\description{
Generate Violin Plot
}
