% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize_data}
\alias{normalize_data}
\title{Normalize and Scale Data for Statistical Analysis}
\usage{
normalize_data(corrected_data, method = "tsn")
}
\arguments{
\item{corrected_data}{matrix of QC-corrected mass spec values with metabolites on the rows and one column per sample.}

\item{method}{The scaling method: "tsn" (Total Sum Normalization, default) or "auto" (Autoscaling).}
}
\value{
The scaled data frame.
}
\description{
Applies a scaling method (e.g., Total Sum Normalization) to metabolite values.
}
