% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_drift.R
\name{qc_drift_correction}
\alias{qc_drift_correction}
\alias{correct_linear}
\title{Correct for QC Drift}
\usage{
qc_drift_correction(raw_data, correction_function = correct_linear, ...)

correct_linear(io, abundance, qc, metabolites)
}
\arguments{
\item{raw_data}{A data frame containing the raw mass spectrometry data with columns for injection order (\code{io}), abundance (\code{abundance}), and a logical column indicating QC samples (\code{qc}).}

\item{correction_function}{A function to apply for drift correction. Currently, the only supported function is \code{correct_linear}, which performs linear interpolation between QC samples.}

\item{io}{A vector indicating injection order}

\item{abundance}{A vector of uncorrected abundances}

\item{qc}{A logical vector indicating QC samples}

\item{metabolites}{A vector identifying each metabolite}
}
\value{
\code{qc_drift_correction} returns a data frame with an additional column, \code{corrected_abundance}, containing the QC-corrected abundances returned by \code{correction_function}.
}
\description{
These functions apply a correction for analytical drift in mass spectrometry data.
They assume that \code{raw_data} has appropriate columns for the chosen correction function.

\code{correct_linear} corrects for analytical drift in mass spectrometry data
by linearly interpolating between QC samples. It assumes that the first and
last samples are QC samples.
}
