% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_and_clean_data}
\alias{read_and_clean_data}
\title{Read and Clean Mass Spectrometry Data}
\usage{
read_and_clean_data(filepath, sheet_name = 1, qc_starts_with = "QC")
}
\arguments{
\item{filepath}{Path to the Excel file.}

\item{sheet_name}{The name or index of the sheet to read. Defaults to 1.}

\item{qc_starts_with}{A character string prefix to identify QC samples. Defaults to 'QC'.}
}
\value{
A tibble in long format with columns: \code{metabolites}, \code{sample}, \code{abundance}, \code{io} (injection order), and \code{qc} (a boolean indicating if the sample is a QC sample).
}
\description{
This function reads a sheet from an Excel file containing mass spectrometry data,
handles duplicate metabolite names, and transforms the data into a long format.
It also adds injection order and identifies QC samples.
}
