% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{run_pca}
\alias{run_pca}
\title{Run PCA on Long-Format Data}
\usage{
run_pca(data, vbl)
}
\arguments{
\item{data}{A data.frame with metabolite information - assumed to be generated by \code{process_runs} or \code{read_and_clean_data}}

\item{vbl}{Character defining the column of \code{data} to use for the PCA}
}
\value{
A \code{prcomp} object with the PCA results
}
\description{
This function takes a long-format data frame, pivots it to a wide format,
and then performs Principal Component Analysis (PCA).
}
